py3k: Hajime Nakagami's Python 2.7/3.3 port, as a single commit
This is Hajime Nakagami's port of PIL to Python 2.7 and 3.3. The original
patch series can be found at:

  https://github.com/nakagami/Pillow/tree/py33py27

The particular commit was:

  5913bafccd60db2ea147982dd2f67970b0bc52eb

+++ PIL/WmfImagePlugin.py
@@ -17,7 +17,7 @@
 
 __version__ = "0.2"
 
-import Image, ImageFile
+from PIL import Image, ImageFile
 
 _handler = None
 
@@ -113,15 +113,15 @@ class WmfStubImageFile(ImageFile.StubImageFile):
             if s[22:26] != "\x01\x00\t\x00":
                 raise SyntaxError("Unsupported WMF file format")
 
-        elif long(s) == 1 and s[40:44] == " EMF":
+        elif int(s) == 1 and s[40:44] == " EMF":
             # enhanced metafile
 
             # get bounding box
-            x0 = long(s, 8); y0 = long(s, 12)
-            x1 = long(s, 16); y1 = long(s, 20)
+            x0 = int(s, 8); y0 = int(s, 12)
+            x1 = int(s, 16); y1 = int(s, 20)
 
             # get frame (in 0.01 millimeter units)
-            frame = long(s, 24), long(s, 28), long(s, 32), long(s, 36)
+            frame = int(s, 24), int(s, 28), int(s, 32), int(s, 36)
 
             # normalize size to 72 dots per inch
             size = x1 - x0, y1 - y0
